# Presentation/wwwroot/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WebStore API",
    "description": "API для заказа товаров из интернет-магазина.\n\nРеализована аутентификация через куки, распределение прав доступа по ролям: гость, пользователь, администратор (помимо своих прав пользователь имеет права гостя, администратор - права гостя и пользователя). Авторизованный пользователь может использовать в запросах только свой логин, администратор - логины всех пользователей.\n\nАдминистратор:\n\nлогин - admin,\n\nпароль - admin.\n\n",
    "version": "1.0.0"
  },
  "paths": {
    "/register": {
      "post": {
        "tags": ["User"],
        "summary": "Регистрация нового пользователя",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "UserName": "john_doe",
                  "Password": "12345",
                  "FirstName": "John",
                  "LastName": "Doe",
                  "PhoneNumber": "+1234567890",
                  "Email": "john@mail.ru"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Пользователь успешно зарегистрирован"
          },
          "400": {
            "description": "Некорректное тело запроса"
          },
          "409": {
            "description": "Логин занят"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": ["User"],
        "summary": "Вход в аккаунт",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешный вход в аккаунт"
          },
          "401": {
            "description": "Неверные учетные данные"
          }
        }
      }
    },
    "/logout": {
      "get": {
        "tags": ["User"],
        "summary": "Выход из аккаунта",
        "responses": {
          "200": {
            "description": "Произведен выход из аккаунта"
          },
          "405": {
            "description": "Пользователь не авторизован"
          }
        }
      }
    },
    "/users/{username}": {
      "get": {
        "tags": ["User"],
        "summary": "Получение информации о пользователе",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о пользователе",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Пользователь не найден"
          }
        }
      },
      "patch": {
        "tags": ["User"],
        "summary": "Назначение прав администратора пользователю",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAdmin",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Права администратора успешно назначены"
          },
          "404": {
            "description": "Пользователь не найден"
          },
          "409": {
            "description": "Пользователь уже является администратором"
          }
        }
      },
      "delete": {
        "tags": ["User"],
        "summary": "Удаление пользователя вместе с информацией о его заказах",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Пользователь удален"
          },
          "404": {
            "description": "Пользователь не найден"
          },
          "405": {
            "description": "Доступ запрещен"
          }
        }
      }
    },
    "/orders": {
      "get": {
        "tags": ["Order"],
        "summary": "Получение списка заказов пользователя",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список заказов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Пользователь не найден"
          }
        }
      },
      "post": {
        "tags": ["Order"],
        "summary": "Оформление заказа",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "UserName": "name",
                  "OrderedGoods": [
                    {
                      "goodId": 1001,
                      "amount": 0
                    },
                    {
                      "goodId": 1002,
                      "amount": 1
                    },
                    {
                      "goodId": 1003,
                      "amount": 2
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Заказ создан"
          },
          "400": {
            "description": "Ошибка в теле запроса"
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Пользователь не найден"
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": ["User"],
        "summary": "Получение списка пользователей",
        "responses": {
          "200": {
            "description": "Список пользователей",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "405": {
            "description": "Доступ запрещен"
          }
        }
      }
    },

    "/orders/{orderId}": {
      "get": {
        "tags": ["Order"],
        "summary": "Получение информации о заказе по заданному id",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о заказе по заданному id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен. Пользователь не имеет доступа к запрашиваемому ресурсу."
          },
          "404": {
            "description": "Заказ не найден."
          }
        }
      },
      "patch": {
        "tags": ["Order"],
        "summary": "Изменение статуса заказа",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "isDone",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Статус заказа обновлен"
          },
          "404": {
            "description": "Заказ не найден / Пользователь не найден"
          },
          "405": {
            "description": "Доступ запрещен"
          }
        }
      }
    },
    
    "/orderedGoods": {
      "get": {
        "tags": ["OrderedGoods"],
        "summary": "Получение списка товаров по заданному id заказа",
        "parameters": [
          {
            "name": "orderId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список товаров по заданному id заказа",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderedGood"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Заказ не найден / Пользователь не найден"
          }
        }
      }
    },
    
    "/goods": {
      "get": {
        "tags": ["Goods"],
        "summary": "Получение списка товаров",
        "responses": {
          "200": {
            "description": "Список товаров",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Good"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "405": {
            "description": "Доступ запрещен"
          }
        }
      }
    }
  },
  
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "UserName": {
            "type": "string"
          },
          "FirstName": {
            "type": "string"
          },
          "LastName": {
            "type": "string"
          },
          "PhoneNumber": {
            "type": "string"
          },
          "Email": {
            "type": "string"
          },
          "IsAdmin": {
            "type": "boolean"
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer"
          },
          "UserName": {
            "type": "string"
          },
          "IsDone": {
            "type": "boolean"
          }
        }
      },
      "OrderedGood": {
        "type": "object",
        "properties": {
          "OrderId": {
            "type": "integer"
          },
          "GoodId": {
            "type": "integer"
          },
          "Amount": {
            "type": "integer"
          },
          "Name": {
            "type": "string"
          },
          "Price": {
            "type": "number"
          }
        }
      },
      "Good": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer"
          },
          "Name": {
            "type": "string"
          },
          "Price": {
            "type": "number"
          }
        }
      }
    }
  }
}